Fix tests: move test functions to proper test_sample.py file
